Estimator_for_prediction_surviving_in_Titanic
This class was made for compiling and fitting the keras model for prediction Surviving in Titanic - clasification exercise. Dataset was got from Kaggle. 
P.S. Dataloader is the class for data-preprocessing of Titanic dataset: filling nans, dropping, working with categorical data
P.S.1. SAVED_ESTIMATOR is a path where model is saved in .pickle format (it's needed for task in courses)